node/node-agent/requirements.txt
aiohttp>=3.8.0
requests>=2.28.0
sentence_transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
textual>=0.44.0
rich>=13.0.0
python-dotenv>=1.0.0

